requirements.txt
img2pdf==0.4.4
lxml==4.8.0
packaging==21.3
pikepdf==5.1.3
Pillow==9.1.1
pyparsing==3.0.9
PyPDF2==1.27.12
